move global function mgr from node mgr to global state mgr
