+++ fe/fe-core/src/main/java/com/starrocks/server/NodeMgr.java
@@ -39,8 +39,6 @@ import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
 import com.starrocks.catalog.BrokerMgr;
 import com.starrocks.catalog.FsBroker;
-import com.starrocks.catalog.Function;
-import com.starrocks.catalog.GlobalFunctionMgr;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.Config;
 import com.starrocks.common.ConfigBase;
@@ -125,7 +123,6 @@ public class NodeMgr {
     private final SystemInfoService systemInfo;
     private final BrokerMgr brokerMgr;
     private final HeartbeatMgr heartbeatMgr;
-    private final GlobalFunctionMgr globalFunctionMgr;
 
     private final GlobalStateMgr stateMgr;
 
@@ -138,7 +135,6 @@ public class NodeMgr {
         this.heartbeatMgr = new HeartbeatMgr(systemInfo, !isCheckpoint);
         this.brokerMgr = new BrokerMgr();
         this.stateMgr = globalStateMgr;
-        this.globalFunctionMgr = new GlobalFunctionMgr();
     }
 
     public void initialize(String[] args) throws Exception {
@@ -198,10 +194,6 @@ public class NodeMgr {
 
     public BrokerMgr getBrokerMgr() {
         return brokerMgr;
-    }
-
-    public GlobalFunctionMgr getGlobalFunctionMgr() {
-        return globalFunctionMgr;
     }
 
     public void getClusterIdAndRoleOnStartup() throws IOException {
@@ -1110,28 +1102,6 @@ public class NodeMgr {
             }
         }
 
-        return checksum;
-    }
-
-    public long loadGlobalFunctions(DataInputStream dis, long checksum) throws IOException {
-        int count = dis.readInt();
-        checksum ^= count;
-        for (long i = 0; i < count; ++i) {
-            Function f = Function.read(dis);
-            globalFunctionMgr.replayAddFunction(f);
-        }
-        return checksum;
-    }
-
-    public long saveGlobalFunctions(DataOutputStream dos, long checksum) throws IOException {
-        List<Function> functions = globalFunctionMgr.getFunctions();
-        int size = functions.size();
-        checksum ^= size;
-        dos.writeInt(size);
-
-        for (Function f : functions) {
-            f.write(dos);
-        }
         return checksum;
     }
